Update hook and README
